Revert "Trying to ignore the rest of the tests to see if it has an effect"
This reverts commit c28f2d1f6f3ce51e9f3a9d368404c1c0e1fba713.
